+++ src/global/normalise.js
@@ -1,3 +1,11 @@
+/*global
+    $,
+    interpretString,
+    IS_PROXY_AVAILABLE,
+    identity,
+    memoise
+*/
+
 /**
  * Normalises a WAI-ARIA attribute name so that it's always lower case and
  * always stars with <code>aria-</code>. If the unprefixed value appears in
@@ -52,9 +60,11 @@
 var normalise = memoise(
     function (name) {
 
+        "use strict";
+
         var prefix = "aria-";
         var lower = interpretString(name).toLowerCase();
-        var full = startsWith.call(lower, prefix)
+        var full = (/^aria-/).test(lower)
             ? lower
             : prefix + lower;
         var stem = full.slice(prefix.length);
@@ -73,6 +83,10 @@ var normalise = memoise(
     IS_PROXY_AVAILABLE
         ? identity
         : function (name) {
+
+            "use strict";
+
             return name + "|" + JSON.stringify($.ariaFix);
+
         }
 );